chore: move oauth to insights_mcp package
